+++ compose.yml
@@ -47,7 +47,6 @@ services:
       - node_modules_ghost_email-addresses:/home/ghost/ghost/email-addresses/node_modules:delegated
       - node_modules_ghost_email-content-generator:/home/ghost/ghost/email-content-generator/node_modules:delegated
       - node_modules_ghost_email-events:/home/ghost/ghost/email-events/node_modules:delegated
-      - node_modules_ghost_email-service:/home/ghost/ghost/email-service/node_modules:delegated
       - node_modules_ghost_email-suppression-list:/home/ghost/ghost/email-suppression-list/node_modules:delegated
       - node_modules_ghost_express-dynamic-redirects:/home/ghost/ghost/express-dynamic-redirects/node_modules:delegated
       - node_modules_ghost_ghost:/home/ghost/ghost/ghost/node_modules:delegated
@@ -191,7 +190,6 @@ volumes:
   node_modules_ghost_email-addresses: {}
   node_modules_ghost_email-content-generator: {}
   node_modules_ghost_email-events: {}
-  node_modules_ghost_email-service: {}
   node_modules_ghost_email-suppression-list: {}
   node_modules_ghost_express-dynamic-redirects: {}
   node_modules_ghost_ghost: {}
