+++ ghost/core/.c8rc.e2e.json
@@ -8,7 +8,7 @@
   ],
   "reportsDir": "./coverage-e2e",
   "statements": 79,
-  "branches": 82,
+  "branches": 81,
   "functions": 80,
   "lines": 79,
   "include": [
