+++ ghost/core/core/server/services/email-analytics/EmailAnalyticsServiceWrapper.js
@@ -7,7 +7,8 @@ class EmailAnalyticsServiceWrapper {
         }
 
         const EmailAnalyticsService = require('./EmailAnalyticsService');
-        const {EmailEventStorage, EmailEventProcessor} = require('@tryghost/email-service');
+        const EmailEventStorage = require('../email-service/EmailEventStorage');
+        const EmailEventProcessor = require('../email-service/EmailEventProcessor');
         const MailgunProvider = require('./EmailAnalyticsProviderMailgun');
         const {EmailRecipientFailure, EmailSpamComplaintEvent, Email} = require('../../models');
         const StartEmailAnalyticsJobEvent = require('./events/StartEmailAnalyticsJobEvent');
