+++ ghost/core/core/server/services/email-service/EmailEventProcessor.js
