+++ ghost/core/core/server/services/email-service/email-templates/partials/styles.hbs
