+++ ghost/core/core/server/services/email-service/email-templates/template.hbs
