+++ ghost/core/package.json
@@ -83,7 +83,6 @@
     "@tryghost/email-addresses": "0.0.0",
     "@tryghost/email-content-generator": "0.0.0",
     "@tryghost/email-mock-receiver": "0.3.8",
-    "@tryghost/email-service": "0.0.0",
     "@tryghost/email-suppression-list": "0.0.0",
     "@tryghost/errors": "1.3.5",
     "@tryghost/express-dynamic-redirects": "0.0.0",
@@ -244,6 +243,7 @@
     "expect": "29.3.1",
     "find-root": "1.1.0",
     "form-data": "4.0.0",
+    "html-validate": "8.29.0",
     "inquirer": "8.2.6",
     "jwks-rsa": "3.2.0",
     "mocha": "10.8.2",
