+++ ghost/core/test/unit/server/services/email-service/email-controller.test.js
@@ -1,5 +1,5 @@
 const assert = require('assert/strict');
-const EmailController = require('../lib/EmailController');
+const EmailController = require('../../../../../core/server/services/email-service/EmailController');
 const {createModel, createModelClass} = require('./utils');
 
 describe('Email Controller', function () {
