--- ghost/email-service/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "@tryghost/email-service",
-  "version": "0.0.0",
-  "repository": "https://github.com/TryGhost/Ghost/tree/main/packages/email-service",
-  "author": "Ghost Foundation",
-  "private": true,
-  "main": "index.js",
-  "scripts": {
-    "dev": "echo \"Implement me!\"",
-    "test:unit": "NODE_ENV=testing c8 --all --check-coverage --100 --reporter text --reporter cobertura -- mocha --reporter dot './test/**/*.test.js'",
-    "test": "yarn test:unit",
-    "lint:code": "eslint *.js lib/ --ext .js --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
-  },
-  "files": [
-    "index.js",
-    "lib"
-  ],
-  "devDependencies": {
-    "c8": "8.0.1",
-    "html-validate": "8.29.0",
-    "mocha": "10.8.2",
-    "should": "13.2.3",
-    "sinon": "15.2.0"
-  },
-  "dependencies": {
-    "@tryghost/color-utils": "0.2.2",
-    "@tryghost/email-events": "0.0.0",
-    "@tryghost/errors": "1.3.5",
-    "@tryghost/html-to-plaintext": "0.0.0",
-    "@tryghost/kg-default-cards": "10.1.1",
-    "@tryghost/logging": "2.4.21",
-    "@tryghost/tpl": "0.1.32",
-    "@tryghost/validator": "0.2.14",
-    "bson-objectid": "2.0.4",
-    "cheerio": "0.22.0",
-    "handlebars": "4.7.8",
-    "juice": "9.1.0",
-    "moment-timezone": "0.5.45"
-  }
-}
